Avoid assuming that quoted posts are stapled
The Bsky API no longer staples quoted post when a user blocks the
quoting user; this is a strange decision because the post is still
quoted (the cid of the post is still attached), meaning that you can
ignore the block and fetch the original content. However, we will just
provide a link to it rather than go to the trouble of fetching the post
directly.